mqtt_pal_recvall: break the loop after reading the requested amount of data
Instead of asking the underlying recv()-like api to read zero bytes.
This simplifies the handling of the return value of the recv()-like
functions and fixes an actual bug at least in the unix socket version.
(TEST__api__publish_subscribe__multiple test case was failing because of
this.)

I tested the unix socket one with TEST__api__publish_subscribe__multiple.
I build-tested the mbedtls one.
Others are totally untested.